+++ test-plugins/org.yakindu.sct.generator.java.test/src-gen/org/yakindu/scr/guard/GuardStatemachine.java
@@ -22,13 +22,13 @@ public class GuardStatemachine implements IGuardStatemachine {
 			return_ID = true;
 		}
 
-		private int myVar;
+		private long myVar;
 
-		public int getMyVar() {
+		public long getMyVar() {
 			return myVar;
 		}
 
-		public void setMyVar(int value) {
+		public void setMyVar(long value) {
 			this.myVar = value;
 		}
 
@@ -92,14 +92,23 @@ public class GuardStatemachine implements IGuardStatemachine {
 		exitAction();
 	}
 
+	/**
+	 * This method resets the incoming events (time events included).
+	 */
 	protected void clearEvents() {
 		sCInterface.clearEvents();
 
 	}
 
+	/**
+	 * This method resets the outgoing events.
+	 */
 	protected void clearOutEvents() {
 	}
 
+	/**
+	 * Returns true if the given state is currently active otherwise false.
+	 */
 	public boolean isStateActive(State state) {
 		switch (state) {
 			case main_region_A :
@@ -125,11 +134,11 @@ public class GuardStatemachine implements IGuardStatemachine {
 		sCInterface.raiseReturn();
 	}
 
-	public int getMyVar() {
+	public long getMyVar() {
 		return sCInterface.getMyVar();
 	}
 
-	public void setMyVar(int value) {
+	public void setMyVar(long value) {
 		sCInterface.setMyVar(value);
 	}
