+++ test-plugins/org.yakindu.sct.generator.java.test/src-gen/org/yakindu/scr/parenthesis/IParenthesisStatemachine.java
@@ -4,8 +4,8 @@ import org.yakindu.scr.IStatemachine;
 public interface IParenthesisStatemachine extends IStatemachine {
 
 	public interface SCInterface {
-		public int getErg();
-		public void setErg(int value);
+		public long getErg();
+		public void setErg(long value);
 
 	}
